Rename test units
