+++ tests/test_marvinLog.py
